pyproject.toml
[project]
name = "visa-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "google-genai>=1.49.0",
    "ipykernel>=7.1.0",
    "mcp[cli]>=1.20.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
]
